chore: move to edgegateway category and minor fix
