+++ .changelog/340.txt
@@ -1,7 +1,7 @@
 ```release-note:new-resource
-`resource/cloudavenue_network_firewall` - New resource to create a network firewall.
+`resource/cloudavenue_edgegateway_firewall` - New resource to create a Edge Gateway firewall.
 ```
 
 ```release-note:new-datasource
-`datasource/cloudavenue_network_firewall` - New datasource to get information about a network firewall.
+`datasource/cloudavenue_edgegateway_firewall` - New datasource to get information about a Edge Gateway firewall.
 ```
