+++ docs/data-sources/edgegateway_firewall.md
@@ -1,23 +1,15 @@
 ---
-page_title: "cloudavenue_network_firewall Data Source - cloudavenue"
-subcategory: "Network"
+page_title: "cloudavenue_edgegateway_firewall Data Source - cloudavenue"
+subcategory: "Edge Gateway (Tier-1)"
 description: |-
   The firewall data source allows you to retrieve information about an Firewall.
 ---
 
-# cloudavenue_network_firewall (Data Source)
+# cloudavenue_edgegateway_firewall (Data Source)
 
 The firewall data source allows you to retrieve information about an Firewall.
 
-## Example Usage
 
-```terraform
-data "cloudavenue_edgegateways" "example" {}
-
-data "cloudavenue_network_firewall" "example" {
-  edge_gateway_id = data.cloudavenue_edgegateways.example.edge_gateways[0].id
-}
-```
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
@@ -29,7 +21,7 @@ data "cloudavenue_network_firewall" "example" {
 
 ### Read-Only
 
-- `id` (String) The ID of the network.
+- `id` (String) The ID of the Firewall Edge Gateway Service.
 - `rules` (Attributes List) The list of rules to apply to the firewall. (see [below for nested schema](#nestedatt--rules))
 
 <a id="nestedatt--rules"></a>
