+++ docs/resources/edgegateway_firewall.md
@@ -1,32 +1,15 @@
 ---
-page_title: "cloudavenue_network_firewall Resource - cloudavenue"
-subcategory: "Network"
+page_title: "cloudavenue_edgegateway_firewall Resource - cloudavenue"
+subcategory: "Edge Gateway (Tier-1)"
 description: |-
   The firewall resource allows you to manage rules on an Firewall.
 ---
 
-# cloudavenue_network_firewall (Resource)
+# cloudavenue_edgegateway_firewall (Resource)
 
 The firewall resource allows you to manage rules on an Firewall.
 
-## Example Usage
 
-```terraform
-data "cloudavenue_edgegateways" "example" {}
-
-resource "cloudavenue_network_firewall" "example" {
-
-  edge_gateway_id = data.cloudavenue_edgegateways.example.edge_gateways[0].id
-  rules = [
-    {
-      action      = "ALLOW"
-      name        = "allow all IPv4 traffic"
-      direction   = "IN_OUT"
-      ip_protocol = "IPV4"
-    }
-  ]
-}
-```
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
@@ -42,7 +25,7 @@ resource "cloudavenue_network_firewall" "example" {
 
 ### Read-Only
 
-- `id` (String) The ID of the network.
+- `id` (String) The ID of the Firewall Edge Gateway Service.
 
 <a id="nestedatt--rules"></a>
 ### Nested Schema for `rules`
