+++ internal/provider/provider.go
@@ -71,6 +71,7 @@ func (p *cloudavenueProvider) DataSources(_ context.Context) []func() datasource
 		// EDGE GATEWAY
 		edgegw.NewEdgeGatewayDataSource,
 		edgegw.NewEdgeGatewaysDataSource,
+		edgegw.NewFirewallDataSource,
 
 		// VDC
 		vdc.NewVDCsDataSource,
@@ -81,6 +82,7 @@ func (p *cloudavenueProvider) DataSources(_ context.Context) []func() datasource
 		vapp.NewVappDataSource,
 		vapp.NewOrgNetworkDataSource,
 		vapp.NewIsolatedNetworkDataSource,
+
 		// CATALOG
 		catalog.NewCatalogsDataSource,
 		catalog.NewCatalogDataSource,
@@ -99,7 +101,6 @@ func (p *cloudavenueProvider) DataSources(_ context.Context) []func() datasource
 		// NETWORK
 		network.NewNetworkIsolatedDataSource,
 		network.NewNetworkRoutedDataSource,
-		network.NewFirewallDataSource,
 
 		// STORAGE
 		storage.NewProfileDataSource,
@@ -114,6 +115,7 @@ func (p *cloudavenueProvider) Resources(_ context.Context) []func() resource.Res
 
 		// EDGE GATEWAY
 		edgegw.NewEdgeGatewayResource,
+		edgegw.NewFirewallResource,
 
 		// VDC
 		vdc.NewVDCResource,
@@ -151,7 +153,6 @@ func (p *cloudavenueProvider) Resources(_ context.Context) []func() resource.Res
 		network.NewNetworkRoutedResource,
 		network.NewNetworkIsolatedResource,
 		network.NewPortProfilesResource,
-		network.NewFirewallResource,
 	}
 }
