+++ templates/resources/edgegateway_firewall.md.tmpl
@@ -1,6 +1,6 @@
 ---
 page_title: "{{.Name}} {{.Type}} - {{.ProviderName}}"
-subcategory: "Network"
+subcategory: "Edge Gateway (Tier-1)"
 description: |-
 {{ .Description | plainmarkdown | trimspace | prefixlines "  " }}
 ---
